pyproject.toml
[build-system]
requires = ["hatchling>=1.25", "hatch-vcs>=0.3"]
build-backend = "hatchling.build"

[project]
name = "emailtoolkit"
dynamic = ["version"]
description = "RFC-aware email parsing, normalization, extraction, and DNS health checks with env-config and a phonenumbers-like API."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "ItsYourBoyRoy", email = "roy.dawson.iv@gmail.com" }]
keywords = ["email", "validation", "dns", "parser", "normalizer", "idna", "rfc", "email-validator"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = ["email-validator>=2.2.0", "idna>=3.10"]

[project.optional-dependencies]
dns = ["dnspython>=2.7.0"]
dotenv = ["python-dotenv>=1.1.1"]
dev = [
    "pytest>=8.2",
    "pytest-rerunfailures>=14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "build>=1.2.0",
    "twine>=5.1.0",
]
[project.scripts]
emailtoolkit = "emailtoolkit.main:_cli"

[tool.hatch.version]
source = "vcs"
fallback-version = "0.0.0"

[tool.hatch.build.hooks.vcs]
version-file = "src/emailtoolkit/_version.py"

[tool.hatch.build]
include = [
  "README.md",
  "LICENSE",
  "THIRD_PARTY_NOTICES.md",
  "src/emailtoolkit/**",
]

[tool.hatch.build.targets.wheel]
packages = ["src/emailtoolkit"]

[tool.hatch.build.targets.editable]
packages = ["src/emailtoolkit"]

[project.urls]
Homepage = "https://github.com/imyourboyroy/emailtoolkit"
Repository = "https://github.com/imyourboyroy/emailtoolkit"
"Bug Tracker" = "https://github.com/imyourboyroy/emailtoolkit/issues"

[tool.pytest.ini_options]
addopts = "-q -p no:capture"
testpaths = ["tests"]

